feat: Add phone number utility function
This commit introduces a new utility file `dial_code_ny_country_code.dart` and a function `getFullPhoneNumber`.

This function constructs a full phone number, including the dial code, from a given country code and phone number. It retrieves the dial code by looking up the country code in a predefined list. If the country code or phone number is missing, or if the country cannot be found, it returns a default string 'User NOT LOGGED'.